refactor to remove Poseidon2Params, drop ~25Kb to 148.4
