--- packages/ultrahonk/src/gadgets/poseidon2/mod.rs
@@ -1,5 +1,0 @@
-pub mod poseidon2_bn254_t4;
-pub mod poseidon2_params;
-
-pub use poseidon2_bn254_t4::POSEIDON2_BN254_T4_PARAMS;
-pub use poseidon2_params::Poseidon2Params;
