+++ packages/ultrahonk/src/lib.rs
@@ -8,7 +8,6 @@ pub mod backends;
 pub mod constants;
 pub mod crs;
 pub mod decider;
-pub mod gadgets;
 pub mod keys;
 pub mod oink;
 pub mod serialize;
@@ -74,5 +73,12 @@ impl Utils {
 
         // Now value_bytes is the 256-bit little-endian representation
         Fq::from_le_bytes_mod_order(&value_bytes)
+    }
+
+    /// Reads a field elemnent from a hexadecimal string. Therebey, the format can or can not include the 0x prefix, i.e., "0x2" and "2" give the same result.
+    fn field_from_hex_string(str: &str) -> Result<ScalarField, &'static str> {
+        let s = str.strip_prefix("0x").unwrap_or(str);
+        let bytes = hex::decode(s).unwrap();
+        Ok(ScalarField::from_be_bytes_mod_order(&bytes))
     }
 }
